requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.1
pymongo==4.5.0
motor==3.3.2
cryptography==43.0.0
httpx==0.27.2
aiohttp==3.9.5
pydantic==2.9.0
pydantic-settings==2.5.0
fastapi==0.115.0
uvicorn==0.30.0
gunicorn==23.0.0
apscheduler==3.10.4
pytz==2024.1
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.2
